Added permissions to delete item

+++ backend/src/resolvers/Mutation.js
@@ -47,9 +47,17 @@ const Mutations = {
 	async deleteItem(parent, args, ctx, info) {
 		const where = { id: args.id };
 		// 1. find the item
-		const item = await ctx.db.query.item({ where }, `{ id title}`);
+		const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+		// TODO['DOG', 'CAT'].some(pet=>['BIRD','SNAKE'].includes(pet))return false
+		//['DOG', 'CAT'].some(pet=>['BIRD','SNAKE','CAT'].includes(pet))return true
 		// 2. Check if they own that item, or have the permissions
-		// TODO
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = ctx.request.user.permissions.some((permission) =>
+			[ 'ADMIN', 'ITEMDELETE' ].includes(permission)
+		);
+		if (!ownsItem && !hasPermissions) {
+			throw new Error("You don't have permissions to do that!");
+		}
 		// 3. Delete it!
 		return ctx.db.mutation.deleteItem({ where }, info);
 	},
